Don't block on conversion endpoints
Getting a conversion always makes a DB request, which is IO-bound, so
async increases request throughput.

Performing a conversion sometimes makes a DB request (IO-bound), and is
an operation of potentially highly variable resource demand, so it is
best not to block so that resource-light requests aren't delayed by
resource-intensive requests.