Re-register SQLAlchemy Extension With Flask App
During the refactor, to make dependency injection work, the db code was
rewritten so that its sessions are no longer managed by the Flask app.
Having the database object provided by the dependency injection
framework makes it easy to swap it out with a mock for testing.
However, it was simpler when Flask automatically managed the db
sessions.

This commit keeps the benefits of both using a DI framework and having
Flask manage the SQLAlchemy.